# secp256k1: Expose Jacobian point equivalency func.