… dcrec/secp256k1/bench_test.go
… dcrec/secp256k1/pubkey_bench_test.go