This exposes a new function on the `JacobianPoint` type named `EquivalentNonConst` which efficiently determines if two Jacobian points represent the same affine point without actually converting the points to affine.

This provides a significant speedup versus first converting to affine for use cases that need the functionality.  One example where it is useful is adaptor signatures.

It includes comprehensive tests for edge conditions as well as ongoing randomized testing.

Finally, it also includes some preliminary commits to make the benchmark structure more consistent and introduce a new benchmark for testing Jacobian point equivalency.

The following benchmark shows a before and after comparison of checking Jacobian point equivalency with the new method versus the affine conversion approach:

```
name                       old time/op   new time/op  delta
--------------------------------------------------------------------------------
JacobianPointEquivalency   17.2µs ± 2%   0.5µs ± 1%   -97.24%  (p=0.000 n=10+10)
```