@davecgh 2024-09-05
> LGTM. NBD, but should it be called `EquivalentNonConst`?

Good point.  That would be less likely to potentially cause confusion of exact equality versus equivalent points.  Updated.